revert changes to auth
